Ontology-based-RAG/backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# LightRAG (GraphRAG implementation)
lightrag-hku>=0.0.3
# Or install from git: git+https://github.com/HKUDS/LightRAG.git

# OpenAI for embeddings (required by LightRAG)
openai>=1.0.0

# HTTP client
httpx>=0.27.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# CORS
fastapi-cors>=0.0.6

# Caching
redis>=5.0.0

# Logging
loguru>=0.7.0
